requirements.txt
# local package
-e .

# external requirements
numpy
pandas_datareader
matplotlib

